Initial implementation

+++ src/resolvers/Mutation.js
@@ -64,11 +64,14 @@ async function updateNote(parent, { titleId, body }, context){
     });
 
     return note;
-
-async function editNote(parent, {titleId}, context){
 }
 
-async function deleteNote(parent, {titleId}, context){
+async function deleteNote(parent, { titleId }, context){
+    const userId = getUserId(context);
+    const { id: noteUserId } = await context.prisma.note({ titleId }).createdBy();
+
+    if (userId != noteUserId)
+        throw new Error("Invalid user! Change this message");
 
     const note = await context.prisma.deleteNote({
         titleId
@@ -82,6 +85,5 @@ module.exports = {
     login,
     createNote,
     updateNote,
-  //  editNote,
     deleteNote
 };
